moves hooks to hooks
